final-pjt-back/requirements.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.8.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.6
django-allauth==0.54.0
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
drf-spectacular==0.27.2
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
fonttools==4.50.0
fqdn==1.5.1
gevent==24.2.1
geventhttpclient==2.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.28.0
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.4
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
locust==2.26.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack==1.0.8
nbclient==0.9.0
nbconvert==7.14.1
nbformat==5.9.2
nest-asyncio==1.5.9
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roundrobin==0.0.4
rpds-py==0.17.1
selenium==4.19.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
trio==0.25.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
# typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.0.4
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
wsproto==1.2.0
zipp==3.18.1
zope.event==5.0
zope.interface==6.3
